# src/main/java/io/swagger/api/CompanyApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.9).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Companies;
import io.swagger.model.Company;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-11-03T01:43:53.014-06:00")

@Api(value = "company", description = "the company API")
public interface CompanyApi {

    @ApiOperation(value = "Get a company by its ID", nickname = "companyCompanyIDGet", notes = "Numeric ID of Driver", response = Company.class, authorizations = {
        @Authorization(value = "APIKeyHeader")
    }, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Company.class) })
    @RequestMapping(value = "/company/{companyID}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Company> companyCompanyIDGet(@ApiParam(value = "Numeric ID of the user to get",required=true) @PathVariable("companyID") Integer companyID);


    @ApiOperation(value = "say Hi", nickname = "companyGet", notes = "", response = Companies.class, authorizations = {
        @Authorization(value = "APIKeyHeader")
    }, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "get all the companies on successful execution", response = Companies.class) })
    @RequestMapping(value = "/company",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Companies> companyGet();

}